Сжать активное ядро: удалить archive/, мёртвый код и дублирующие тестовые модули
