--- .gitkeep
@@ -1,1 +1,0 @@
-# .gitkeep file auto-generated at 2026-07-20T13:49:21.880Z for PR creation at branch issue-65-b0e2f91ea39b for issue https://github.com/netkeep80/anum_docs/issues/65
