--- archive/2.md
@@ -1,173 +1,0 @@
-## Система аксиом МТС (чистовой вариант)
-
-Учитывая ваши требования и критический пересмотр операторов ♂/♀, предлагаю согласованную систему аксиом в логической последовательности от фундаментальных к производным. Ключевое уточнение: **оператор формальной нотации МТС** — это символ, вводимый аксиомой определения и обозначающий конкретную форму связи или преобразование форм.
-
----
-
-### 📌 Фундаментальные аксиомы
-
-#### Аксиома существования связи
-**Формула:**
-```
-rv : (r ⟼ v)
-```
-**Содержательно-онтологическое определение:**  
-Существование есть направленная связь от начала к концу.  
-**Структурные свойства:**  
-`⟼` — бинарный, прямоассоциативный оператор формальной нотации МТС:  
-`abc ≡ (ab)c ≢ a(bc)`  
-**Запрет:** Запрещена транзитивность: `(a ⟼ b) ⟼ c ≢ a ⟼ (b ⟼ c)`.
-
----
-
-#### Аксиома смысла ∞
-**Формула:**
-```
-∞ : (∞ ⟼ ∞)
-```
-**Содержательно-онтологическое определение:**  
-Смысл есть связь смысла со смыслом (полное самозамыкание).  
-**Структурные свойства:**  
-`∞` — нульарный оператор формальной нотации МТС, единственная полностью самозамкнутая связь.  
-**Следствие:** `∞` служит базисом сериализации: каждое ачисло неявно начинается с `∞`.
-
----
-
-### 🔁 Аксиомы извлечения начала и конца
-
-#### Аксиома извлечения конца связи (оператор ♂)
-**Формула:**
-```
-♂(a ⟼ b) = b
-```
-**Содержательно-онтологическое определение:**  
-Безначальная связь есть конец связи. Оператор `♂` извлекает смысл конца из формы связи.  
-**Структурные свойства:**  
-`♂` — унарный постфиксный оператор формальной нотации МТС.  
-**Остенсивное определение:** Стрелка с началом, замкнутым на её собственное тело — указывает на отсутствие начала, наличие только конца.
-
----
-
-#### Аксиома извлечения начала связи (оператор ♀)
-**Формула:**
-```
-(a ⟼ b)♀ = a
-```
-**Содержательно-онтологическое определение:**  
-Бесконечная связь есть начало связи. Оператор `♀` извлекает смысл начала из формы связи.  
-**Структурные свойства:**  
-`♀` — унарный постфиксный оператор формальной нотации МТС.  
-**Остенсивное определение:** Стрелка с концом, замкнутым на её собственное тело — указывает на отсутствие конца, наличие только начала.
-
----
-
-### 🧭 Производные аксиомы
-
-#### Аксиома начала смысла
-**Формула:**
-```
-( : ♂∞
-```
-**Содержательно-онтологическое определение:**  
-Начало смысла есть безначальная связь акорня.  
-**Следствие:** `( ≡ ♂∞ ⟼ ∞` (абит начала).
-
----
-
-#### Аксиома конца смысла
-**Формула:**
-```
-) : ∞♀
-```
-**Содержательно-онтологическое определение:**  
-Конец смысла есть бесконечная связь акорня.  
-**Следствие:** `) ≡ ∞ ⟼ ∞♀` (абит конца).
-
----
-
-#### Аксиома единицы связи
-**Формула:**
-```
-1 : ♂∞ ⟼ ∞♀
-```
-**Содержательно-онтологическое определение:**  
-Единица связи есть связь начала смысла с концом смысла.  
-**Следствие:** `1` — абит наличия связи (истина).
-
----
-
-#### Аксиома нуля связи
-**Формула:**
-```
-0 : ∞♀ ⟼ ♂∞
-```
-**Содержательно-онтологическое определение:**  
-Ноль связи есть связь конца смысла с началом смысла.  
-**Следствие:** `0` — абит отсутствия связи (ложь).
-
----
-
-#### Аксиома инверсии направления
-**Формула:**
-```
-¬(a ⟼ b) : ♂(a ⟼ b) ⟼ (a ⟼ b)♀
-```
-**Содержательно-онтологическое определение:**  
-Инверсия связи есть связь конца исходной связи с её началом.  
-**Структурные свойства:**  
-`¬` — унарный оператор формальной нотации МТС, инволютивный: `¬¬x = x`.  
-`¬∞ = ∞` (акорень инвариантен).  
-**Следствие:** `¬1 = 0`, `¬0 = 1`.
-
----
-
-#### Аксиома равенства связей по смыслу
-**Формула:**
-```
-(a ⟼ b) = (c ⟼ d) ⇔ {a = c, b = d}
-```
-**Содержательно-онтологическое определение:**  
-Связи равны по смыслу тогда и только тогда, когда равны их начала и концы.  
-**Структурные свойства:**  
-Равенство рекурсивно и структурно прозрачно — нет скрытых свойств за пределами формы связи.
-
----
-
-#### Аксиома петли
-**Формула:**
-```
-aa : (a ⟼ a)
-```
-**Содержательно-онтологическое определение:**  
-Петля есть связь начала и конца одной и той же связи.  
-**Запрет:** `aa ≢ ∞` при `a ≢ ∞` (петля не есть акорень).
-
----
-
-### 📊 Итоговая таблица операторов формальной нотации МТС
-
-| Символ | Тип | Определение | Смысл |
-|--------|-----|-------------|-------|
-| `⟼` | бинарный | `rv : r ⟼ v` | направленная связь |
-| `∞` | нульарный | `∞ : ∞ ⟼ ∞` | полное самозамыкание (смысл) |
-| `♂` | унарный | `♂(a⟼b) = b` | извлечение конца (безначальность) |
-| `♀` | унарный | `(a⟼b)♀ = a` | извлечение начала (бесконечность) |
-| `¬` | унарный | `¬(a⟼b) : b⟼a` | инверсия направления |
-| `:` | мета | `x : F` | аксиома определения символа |
-| `=` | предикат | `(a⟼b)=(c⟼d) ⇔ {a=c,b=d}` | равенство по смыслу |
-| `(` | абит | `( : ♂∞` | начало связи |
-| `)` | абит | `) : ∞♀` | конец связи |
-| `1` | абит | `1 : ♂∞ ⟼ ∞♀` | наличие связи (истина) |
-| `0` | абит | `0 : ∞♀ ⟼ ♂∞` | отсутствие связи (ложь) |
-
----
-
-### 🔑 Ключевые преимущества системы
-
-1. **Устранена коллизия символа `-`**: инверсия теперь обозначается `¬`, абит лжи — `0`.
-2. **Чёткое разделение мета/объектного уровней**: `:` только для введения символов, `=` только для равенства связей.
-3. **Операторы ♂/♀ переосмыслены конструктивно**: не как «самозамыкание», а как **извлечение** начала/конца — это решает проблему бесконечной регрессии.
-4. **Ачисло как идентификатор**: последовательность абитов `( ) 1 0` однозначно кодирует форму связи через её начало/конец/наличие/отсутствие.
-5. **Логическая последовательность**: от базового конструктора `⟼` → к акорню `∞` → к операторам извлечения → к абитам → к логическим операциям.
-
-Система замкнута: все символы формальной нотации МТС имеют свою аксиому, все термины согласованы с вашими требованиями («смысл» вместо «роли», «начало/конец» вместо «ссылка/значение»).
