--- archive/faq/Аксиоматизация натуральных чисел в МТС.md
@@ -1,91 +1,0 @@
-**Аксиоматизация натуральных чисел в МТС через степень петли (Аксиома Пеано-МТС):**  
-В рамках Метатеории связей (МТС) натуральные числа определяются как **длины повторяющихся последовательностей связей**, формализуемые через оператор степени петли \( a^n \).  
-Введём следующие аксиомы:
-
----
-
-### **1. Аксиома базовой длины (ноль)**  
-Существует **базовая длина** \( 0 \), соответствующая ассоциативному корню \( ∞ \):  
-\[  
-0 \equiv ∞.  
-\]  
-**Смысл:**  
-\( ∞ \) — минимальная длина, не содержащая явных связей (полное самозамыкание).
-
----
-
-### **2. Аксиома инкремента (следующая длина)**  
-Для любой связи \( a \) и длины \( n \), существует **следующая длина** \( n+1 \), определяемая как:  
-\[  
-a^{n+1} \equiv a^n \to a.  
-\]  
-**Пример:**  
-- \( a^1 \equiv a \),  
-- \( a^2 \equiv a \to a \),  
-- \( a^3 \equiv (a \to a) \to a \).  
-
----
-
-### **3. Аксиома уникальности длин**  
-Разные длины соответствуют разным степеням:  
-\[  
-a^n \equiv a^m \implies n \equiv m \quad \text{и} \quad a \equiv a.  
-\]  
-**Смысл:**  
-Длина последовательности однозначно определяет степень.
-
----
-
-### **4. Аксиома запрета самозамкнутости для не-∞**  
-Никакая длина \( n \geq 1 \) не может быть эквивалентна \( ∞ \):  
-\[  
-a^n \not\equiv ∞ \quad \text{при } n \geq 1.  
-\]  
-
----
-
-### **5. Аксиома индукции**  
-Если свойство \( P \) выполняется для:  
-1. Базовой длины \( 0 \equiv ∞ \),  
-2. Для любой длины \( n \), из \( P(n) \) следует \( P(n+1) \),  
-то \( P \) выполняется для **всех длин**.  
-
----
-
-### **Связь с оператором степени петли**  
-- **Натуральное число** \( n \) — это длина последовательности \( a^n \).  
-- **Пример чисел:**  
-  - \( 1 \equiv a^1 \equiv a \),  
-  - \( 2 \equiv a^2 \equiv a \to a \),  
-  - \( 3 \equiv a^3 \equiv (a \to a) \to a \).  
-
----
-
-### **Интеграция с ассоциативными числами (ачислами)**  
-Натуральные числа кодируются как **ассоциативные числа** через абиты `1` (наличие связи) и `0` (отсутствие):  
-- \( 1 \equiv ∞1 \),  
-- \( 2 \equiv ∞11 \),  
-- \( 3 \equiv ∞111 \).  
-
----
-
-### **Пример формализации**  
-1. **Базовый случай:**  
-   \[  
-   0 \equiv ∞.  
-   \]  
-2. **Индуктивный шаг:**  
-   \[  
-   1 \equiv a^1 \equiv a, \quad 2 \equiv a^2 \equiv a \to a, \quad 3 \equiv a^3 \equiv (a \to a) \to a.  
-   \]  
-
----
-
-### **Итог**  
-Аксиомы Пеано адаптированы для МТС следующим образом:  
-1. **Ноль** — базовая длина \( ∞ \).  
-2. **Инкремент** — добавление связи \( a \) к предыдущей длине.  
-3. **Уникальность** — каждая длина соответствует уникальной степени.  
-4. **Индукция** — доказательство свойств для всех длин через базовый случай и шаг.  
-
-Таким образом, натуральные числа в МТС — это **степени связей**, где каждое число есть длина их последовательности.
