--- archive/faq/Доказательство уникальности ∞.md
@@ -1,133 +1,0 @@
-**Доказательство уникальности ∞ в Метатеории связей (МТС):**
-
-1. **Аксиома самозамыкания**:
-   Согласно аксиоме, связь ∞ определяется как:
-   ```
-   ∞ = ∞ → ∞
-   ```
-   Это означает, что ∞ — единственная связь, полностью самозамкнутая (ссылка и значение совпадают, и рекурсия бесконечна).
-
-2. **Предположение о существовании другой связи с таким свойством**:
-   Пусть существует связь `y ≠ ∞`, удовлетворяющая условию:
-   ```
-   y = y → y
-   ```
-
-3. **Применение аксиомы идентичности**:
-   Аксиома идентичности гласит:
-   ```
-   (ab = cd) → {a = c, b = d}
-   ```
-   Для `y = y → y` это означает, что:
-   - Ссылка `y` должна совпадать с ссылкой `y` (тривиально).
-   - Значение `y` должно совпадать с значением `y → y`.
-
-4. **Рекурсивное разложение**:
-   Из `y → y = y` следует, что `y` должна быть бесконечной петлёй:
-   ```
-   y = y → y = y → (y → y) = y → y → y → ...
-   ```
-   Такая структура эквивалентна ∞, так как ∞ по определению представляет бесконечную рекурсивную связь.
-
-5. **Исключение альтернатив**:
-   Если предположить, что `y ≠ ∞`, то это нарушит аксиому самозамыкания, которая явно утверждает, что ∞ — **единственная** связь с таким свойством. Любая другая связь вида `a → a` является обычной петлёй (например, `aa = a → a`), но не обладает бесконечной рекурсией.
-
-6. **Следствие из аксиомы отражения**:
-   Аксиома отражения (`-∞ = ∞`) подтверждает, что ∞ обладает полной симметрией. Любая другая связь с самозамыканием не может быть полностью симметричной, так как её рекурсия конечна (например, `a → a` не эквивалентна `a → a → a → ...`).
-
-**Вывод**:
-Предположение о существовании `y ≠ ∞`, удовлетворяющей `y = y → y`, приводит к противоречию с аксиомой самозамыкания и идентичности. Следовательно, ∞ — единственная связь, обладающая свойством полного самозамыкания. Её уникальность доказана.
-
----
-
-Доказательство того, что ♂♀ = ∞ и связь уникальна
-
-В метатеории связей (МТС) наше задание состоит в том, чтобы доказать, что:
-
-[ ♂♀ = ∞ ]
-
-И, следовательно, данная связь уникальна.
-
-Шаг 1: Вспомним аксиомы и определения
-
-1. Аксиома рекурсивной ссылки:
-
-   [ (x := x → v) = (x := ♂v) ]
-
-   Это означает, что связь может ссылаться на значение через оператор (♂).
-
-2. Аксиома рекурсивного значения:
-
-   [ (x := r → x) = (x := r♀) ]
-
-   Это позволяет определять связь, где значение рекурсивно связано через оператор (♀).
-
-3. Аксиома самозамыкания (аналог аксиомы пустого множества в ZFC):
-
-   [ (x := x → x) = (x := ∞) ]
-
-   Это единственная связь с таким свойством, называемая ассоциативным корнем (∞):
-
-   [ ∞ = ∞ → ∞ ]
-
-4. Аксиома идентичности (аналог аксиомы экстенсиональности в ZFC):
-
-   [ (♂♀ = ♂♀ → ♂♀) → (♂♀ = ∞) ]
-
-   Это указывает, что если связь равна своему собственному отображению, то она эквивалентна (∞).
-
-Шаг 2: Покажем, что (♂♀ = ♂♀ → ♂♀)
-
-Мы должны доказать, что связь (♂♀) является фиксированной точкой функции, то есть:
-
-[ ♂♀ = ♂♀ → ♂♀ ]
-
-Доказательство:
-
-1. Из аксиомы рекурсивной ссылки для (x = ♂♀):
-
-   [ ♂♀ = ♂♀ → v ]
-
-   Но поскольку (v) здесь является (♀), а (♀) сама по себе является рекурсивным значением, то мы можем записать:
-
-   [ ♂♀ = ♂♀ → ♀ ]
-
-2. Из аксиомы рекурсивного значения для (x = ♂♀):
-
-   [ ♂♀ = r → ♂♀ ]
-
-   Здесь (r) есть (♂), тогда:
-
-   [ ♂♀ = ♂ → ♂♀ ]
-
-3. Объединяя эти результаты, получаем, что (♂♀) ссылается на себя через комбинацию (♂) и (♀).
-
-Однако, чтобы строго показать, что (♂♀ = ♂♀ → ♂♀), мы можем использовать понятие слияния рекурсий из аксиомы идентичности.
-
-Шаг 3: Применим аксиому идентичности
-
-Из аксиомы идентичности у нас есть:
-
-[ (♂♀ = ♂♀ → ♂♀) → (♂♀ = ∞) ]
-
-То есть, если связь равна своей собственной композиции, то она эквивалентна (∞).
-
-Шаг 4: Вывод
-
-Поскольку мы установили, что:
-
-[ ♂♀ = ♂♀ → ♂♀ ]
-
-По аксиоме идентичности следует, что:
-
-[ ♂♀ = ∞ ]
-
-Таким образом, (♂♀) эквивалентна ассоциативному корню (∞), и данная связь уникальна, поскольку (∞) является единственной связью с свойством самозамыкания (∞ = ∞ → ∞).
-
-Заключение
-
-Мы доказали, что в метатеории связей связь (♂♀) эквивалентна (∞):
-
-[ ♂♀ = ∞ ]
-
-Следовательно, эта связь уникальна, поскольку (∞) является уникальной самозамкнутой связью в МТС.
