--- archive/faq/Обозначения связей и мультиссылок в теории.md
@@ -1,53 +1,0 @@
-### Формальное описание обозначений для связей и мультиссылок
-
-#### 1. **Тождественность (=)**
-Символ **=** означает структурную идентичность связей.  
-**Пример**:  
-\[ rv \equiv r \rightarrow v \]  
-▸ Левая и правая часть — разные записи одной и той же связи.
-
----
-
-#### 2. **Равенство (=) для мультиссылок**
-Символ **=** полностью определяет пучок исходящих связей для мультиссылки.
-**Пример**:  
-\[ \{ x \rightarrow (a \rightarrow b), \, x \rightarrow (c \rightarrow d) \} \equiv (x = \{ a \rightarrow b, \, c \rightarrow d \}) \]  
-▸ Здесь:  
-- **Левая часть**: Связь \( x \) выступает в роли ссылки для двух значений.  
-- **Правая часть**: \( x \) интерпретируется как мультиссылка, играющая роль множества связей.  
-
----
-
-#### 3. **Правила обозначений**
-
-| **Символ** | **Смысл**                                                                 | **Пример**                                  |
-|------------|---------------------------------------------------------------------------|---------------------------------------------|
-| \( \equiv \) | Тождество структур (одинаковые связи в разных нотациях)                 | \( ab \equiv a \rightarrow b \)            |
-| \( = \)      | Контекстное равенство ролей (ссылка ↔ множество значений)               | \( x = \{ a, b \} \equiv \{ x \rightarrow a, x \rightarrow b \} \) |
-| \( \{ \} \)  | Множество связей, связанных одной мультиссылкой                         | \( x \rightarrow \{ a, b \} \)             |
-
----
-
-#### 4. **Визуализация мультиссылок**
-```
-       x
-     /   \
-    ↓     ↓
-(a→b)  (c→d)
-```
-▸ Связь \( x \) является мультиссылкой, указывающей на два значения: \( a \rightarrow b \) и \( c \rightarrow d \).
-
----
-
-#### 5. **Аксиома множественности**
-\[ \{ x \rightarrow L_1, \, x \rightarrow L_2 \} \equiv x = \{ L_1, L_2 \} \]  
-▸ Если связь \( x \) используется как ссылка для \( L_1 \) и \( L_2 \), она эквивалентна множеству \( \{ L_1, L_2 \} \).
-
----
-
-### Итог
-- **=** — для структурного тождества,  
-- **=** — для определения пучка исходящих связей для связи в роли мультиссылки,  
-- **Мультиссылка** — роль связи для объединения множества связей.  
-Пример:  
-\[ \text{Связь } x \text{ как мультиссылка: } x = \{ \text{значение}_1, \text{значение}_2 \} \equiv \{ x \rightarrow \text{значение}_1, \, x \rightarrow \text{значение}_2 \} \]
