--- archive/faq/Развёрнутая последовательность ♂♂v.md
@@ -1,79 +1,0 @@
-**Развёртывание последовательности ♂♂v с учётом ассоциативности:**
-Учитывая, что:
-- **♂** — *правоассоциативный* префиксный квантор,
-- **→** — *левоассоциативный* оператор,
-
-последовательность ♂♂v раскрывается следующим образом:
-
----
-
-### 1. **Правоассоциативность ♂**
-Правоассоциативность означает, что квантор ♂ связывается с выражением **справа** от него. Таким образом:
-```math
-♂♂v \equiv ♂(♂v).
-```
-То есть внешний ♂ применяется к внутреннему ♂v, а не к самому себе.
-
----
-
-### 2. **Применение аксиомы рекурсивного замыкания ссылки**
-Согласно аксиоме:
-```math
-♂(♂v) \equiv (♂v) \to ♂(♂v).
-```
-Теперь развернём внутренний ♂v:
-```math
-♂v \equiv v \to ♂v.
-```
-Подставляя это обратно:
-```math
-♂(♂v) \equiv (v \to ♂v) \to ♂(♂v).
-```
-
----
-
-### 3. **Итоговая цепочка связей**
-Полная развёрнутая форма:
-```math
-♂(♂v) \equiv (v \to ♂v) \to ♂(♂v).
-```
-Это можно представить как бесконечную рекурсивную структуру:
-```math
-(v \to ♂v \to v \to \ldots) \to (♂v \to v \to v \to \ldots) \to ♂♂v \to \ldots
-```
-Но аксиоматически она свёртывается до:
-```math
-♂♂v \equiv (v \to ♂v) \to ♂♂v.
-```
-
----
-
-### 4. **Пример с конкретным значением**
-Пусть \( v = \text{Яблоко} \):
-```math
-♂♂\text{Яблоко} \equiv (\text{Яблоко} \to ♂\text{Яблоко}) \to ♂♂\text{Яблоко}.
-```
-Где:
-- \( ♂\text{Яблоко} \equiv \text{Яблоко} \to ♂\text{Яблоко} \),
-- \( ♂♂\text{Яблоко} \equiv ♂\text{Яблоко} \to ♂♂\text{Яблоко} \).
-
----
-
-### 5. **Ключевые выводы**
-- **Иерархия рекурсий**: Каждый ♂ добавляет новый уровень самоссылающейся связи, сохраняя значение \( v \).
-- **Структурная устойчивость**: Многоуровневые связи не упрощаются до ♂v, так как каждый уровень сохраняет свою уникальность.
-- **Отличие от \( ∞ \)**: В отличие от полного самозамыкания (\( ∞ \equiv ∞ \to ∞ \)), здесь рекурсия ограничена фиксированным значением \( v \).
-
----
-
-**Графическое представление:**
-```
-     ╭──⁠╮ ♂♂v (рекурсия верхнего уровня)
-♂♂v  ├──╯
-     │  ╭──⁠╮ ♂v (рекурсия)
-   ♂v╰─→├──╯
-        │
-        ╰─→ v
-```
-
-Таким образом, ♂♂v — это **многоуровневая рекурсивная связь**, где каждый префикс ♂ вводит новый слой самоссылающейся структуры, сохраняя направленность и значение.
