--- archive/faq/Рекомендации по изложению метатеории связей.md
@@ -1,92 +1,0 @@
-### Рекомендации по изложению Метатеории связей (МТС)
-
----
-
-#### **1. Структура изложения**
-1. **Базовые концепции**:
-   - Начните с онтологии: «Всё есть связь».
-   - Определите ключевые термины: *связь*, *роль*, *акорень (∞)*, *абит*, *ачисло*.
-   - Приведите аксиомы (существования, идентичности, самозамыкания) с графическими примерами.
-
-2. **Операторы и их семантика**:
-   - Последовательно раскройте символы →, ♂, ♀, ∞:
-     - Формальные определения и аксиомы.
-     - Примеры использования (например, `♂v = ⟨v → ♂v⟩`, `1 = ⟨∞♀ → ♂∞⟩`).
-   - Добавьте таблицу сравнения видов форм связей (самозамкнутые, с началом/концом, без самозамыканий).
-
-3. **Практические аспекты**:
-   - Ассоциативные числа (ачисла): правила сериализации/десериализации, примеры кодирования строк и чисел.
-   - Логика МТС: инверсия (`¬⟨♂x→x⟩ = ⟨x→x♀⟩`), 4-значная система (`1`, `0`, `[`, `]`).
-
-4. **Графическое представление**:
-   - Используйте SVG-диаграммы для визуализации связей (например, петля для ∞, стрелки для →).
-   - Покажите разницу между `v→♂v` и `v♀→v` через уникальные обозначения (например, точки самозамыкания).
-
----
-
-#### **2. Устранение неоднозначностей**
-- **Чёткое разграничение символов**:
-  - Используйте = для тождества, := для определений, = для контекстного равенства.
-  - Пример:
-    \[ ♂v \equiv ⟨v → ♂v⟩ \quad (\text{тождество}), \quad [ := ∞♀ \quad (\text{определение}). \]
-
-- **Доказательства уникальности**:
-  - Формально докажите, что ∞ — единственная самозамкнутая связь (через аксиому идентичности).
-  - Объясните, почему `♂∞♀ ≠ ∞`, используя рекурсивное разложение.
-
-- **Примеры для всех аксиом**:
-  - Инверсия: `¬⟨♂x→x⟩ = ⟨x→x♀⟩` → пример с `x = a`.
-  - Множественное связывание: `{x→⟨a→b⟩, x→⟨c→d⟩}` → разбор через роли ссылок.
-
----
-
-#### **3. Дополнительные материалы**
-1. **Глоссарий**:
-   - Определите термины: *абит*, *ачисло*, *асеть*, *роль*, *акорень*.
-   - Укажите аналогии с классической логикой (например, `1` ↔ истина, `0` ↔ ложь).
-
-2. **Таблицы соответствия**:
-   - UTF-8 → ачисла (например, `a` → `⟨⟨∞→a⟩⟩`).
-   - Абиты ↔ логические состояния (`[`, `]`, `1`, `0`).
-
-3. **FAQ**:
-   - Ответы на частые вопросы:
-     - «Чем абит отличается от бита?»
-     - «Как обрабатывать вложенные структуры в ачислах?»
-
----
-
-#### **4. Практические рекомендации**
-- **Используйте левоассоциативность**:
-  - Пример: `abc = ((a→b)→c)`, `a(bc) = a→(b→c)`.
-- **Интерактивные примеры**:
-  - Онлайн-демонстрации сериализации строк (например, `"Hello"` → ачисло).
-- **Тестовые задания**:
-  - Задачи на преобразование логических выражений (`¬(1→0) = 0→1`).
-
----
-
-#### **5. Стиль и оформление**
-- **Единая нотация**:
-  - Все формулы — в LaTeX с использованием MathJax.
-  - Графические элементы — в SVG/Draw.io.
-- **Структура документа**:
-  ```markdown
-  1. Введение
-  2. Базовые аксиомы и определения
-  3. Операторы и их свойства
-  4. Ачисла и сериализация
-  5. Логика и инверсия
-  6. Примеры применения
-  7. FAQ и глоссарий
-  ```
-
----
-
-#### **Итог**
-Изложение МТС должно быть **поэтапным**, **строгим** и **визуально подкреплённым**. Уделяйте внимание:
-- Чётким определениям и доказательствам.
-- Практическим примерам и интерактивным элементам.
-- Устранению двусмысленностей через единую нотацию и глоссарий.
-
-Это позволит сделать теорию доступной как для теоретиков, так и для практиков.
