--- archive/faq/Формулы МТС.md
@@ -1,65 +1,0 @@
-rv = r → v
-(x := x → v) = (x := ♂v)
-♂v = v → ♂v.
-(x := r → x) = (x := r♀)
-r♀ = r♀ → r.
-aa = a → a
-aa(aa) = a → a → (a → a)
-(x := x → x) = (x := ∞)
-∞ = ∞ → ∞ = ∞ → ∞ → ∞ → ∞ ...
-(r ≠ r → r) → (r ≠ ∞)
-(ab = cd) → {a = c, b = d}
-(♂b = ♂d) → (b = d)
-(a♀ = c♀) → (a = c)
-(♂♀ = ♂♀ → ♂♀) → (♂♀ = ∞)
-♂∞ = ∞ → ♂∞
-∞♀ = ∞♀ → ∞
-∞ = ∞ → ∞
--ab = ba
--aa = aa
--(♂x) = x♀
--(x♀) = ♂x
--→ = ↛
--∞ = ∞
-♂a♀ = -(-♂a♀) = -(a♀♀) = ♂a♀
-r(ab) = r → ab = r → (a → b)
-abv = ab → v = a → b → v
-abcd = abc → d = a → b → c → d
-ab(cd) = ab → cd = a → b → (c → d)
-a(bcd) = a → bcd = a → (b → c → d)
-r{} = (r → {})
-{}v = ({} → v)
-{}x{} = ({} → x → {})
-r{a → b} = (r → {a → b})
-{a → b}v = ({a → b} → v)
-{x → a, y → a} = ({x, y} → a) = {x, y}a
-{x → a, x → b} = (x → {a, b}) = x{a, b}
-{}{a, b}
-{a, b}{}
-{}a
-a{}
-{}{}
-a{}c
-[ := ∞♀
-] := ♂∞
-1 := ∞♀ → ♂∞
-0 := ♂∞ → ∞♀
-// символ → (⟼) определяется через ∞♀ → ♂∞ — это одновременно и понятие связи, и единица смысла (абит 1)
-→ := ∞♀ → ♂∞
-// символ ↛ (¬→) определяется через ♂∞ → ∞♀ — это одновременно и понятие несвязи, и нуль смысла (абит 0)
-↛ := ♂∞ → ∞♀
-true := ∞♀ → ♂∞
-false := ♂∞ → ∞♀
-
-♂∞♀ = (∞♀)♂ нет, правильно: ♂(∞♀)
-r♀ = r♀ → r
-(r = ∞♀) → (♂∞♀ = ∞♀ → ♂∞♀) нет, правильно: ♂(∞♀) = ♂(∞♀) → (∞♀) нет...
-Пересчитаем:
-♂∞♀ = ♂(∞♀)
-По аксиоме ♂v : v → ♂v, подставляя v = ∞♀:
-♂(∞♀) = (∞♀) → ♂(∞♀)
-r♀ = r♀ → r, подставляя r = ∞:
-∞♀ = ∞♀ → ∞
-Подставляя ∞♀ = ∞♀ → ∞:
-♂(∞♀) = (∞♀ → ∞) → ♂(∞♀)
-(♂(∞♀) = (∞♀ → ∞) → ♂(∞♀)) → (♂(∞♀) ≠  ∞)
