--- archive/faq/Функция и множество как бинарное отношение.md
@@ -1,22 +1,0 @@
-Функцию, возвращающую множество, можно описать **обычным бинарным отношением**, но не как функциональное (в стандартном понимании). Разберем детали:
-
-1. **Функция, возвращающая множество**:  
-   Формально это функция вида \( f: A \to \mathcal{P}(B) \), где \( \mathcal{P}(B) \) — множество всех подмножеств \( B \). Для каждого \( a \in A \), значение \( f(a) \) — это подмножество элементов из \( B \).
-
-2. **Бинарное отношение**:  
-   Это подмножество \( R \subseteq A \times B \), где \( (a, b) \in R \) означает, что \( a \) связано с \( b \).
-
-3. **Связь между функцией и отношением**:  
-   Функцию \( f \) можно представить как бинарное отношение \( R \), где:  
-   \[
-   (a, b) \in R \iff b \in f(a).
-   \]  
-   Например, если \( f(a) = \{b_1, b_2\} \), то \( R \) содержит пары \( (a, b_1) \) и \( (a, b_2) \).
-
-4. **Почему это не функциональное отношение**:  
-   - **Функциональное отношение** требует, чтобы для каждого \( a \in A \) существовал **ровно один** \( b \in B \), такой что \( (a, b) \in R \).  
-   - В нашем случае для одного \( a \) может быть **несколько** \( b \), если \( f(a) \) — множество.  
-   - Следовательно, \( R \) не является функциональным отношением (если только \( f(a) \) не одноэлементно для всех \( a \)).
-
-### Итог:
-Да, функцию \( f: A \to \mathcal{P}(B) \) можно описать **бинарным отношением** \( R \subseteq A \times B \), где \( (a, b) \in R \iff b \in f(a) \). Однако это отношение **не будет функциональным** в классическом смысле, так как одному \( a \) может соответствовать множество \( b \).
