--- archive/Аксиоматическое определение начала и конца связи.md
@@ -1,65 +1,0 @@
-# Аксиоматическое определение начала и конца связи в МТС
-
-В Метатеории Связей (МТС) начало и конец связи не являются примитивными понятиями, а определяются рекурсивно через паттерны форм самозамыканий. Вот как это формализуется в системе аксиом:
-
-## Аксиома определения начала связи
-
-```
-♂∞ : (♂∞ ⟼ ∞)
-```
-
-**Содержательное определение:**  
-Начало смысла (♂∞) определяется как связь, в которой начало совпадает с самой связью, а конец указывает на ассоциативный корень. Это паттерн формы, задающий запрос ко всем связям с самозамыканием по началу и фиксированным концом ∞.
-
-**Онтологический статус:**  
-Начало не существует изолированно — оно возникает как роль в акте связывания, когда связь самозамыкается по своему началу и указывает на акорень.
-
-## Аксиома определения конца связи
-
-```
-∞♀ : (∞ ⟼ ∞♀)
-```
-
-**Содержательное определение:**  
-Конец смысла (∞♀) определяется как связь, в которой начало указывает на ассоциативный корень, а конец совпадает с самой связью. Это паттерн формы, задающий запрос ко всем связям с самозамыканием по концу и фиксированным началом ∞.
-
-**Онтологический статус:**  
-Конец не существует изолированно — он возникает как роль в акте связывания, когда связь указывает на акорень и самозамыкается по своему концу.
-
-## Аксиома направленности связи
-
-```
-𝟙 : (♂∞ ⟼ ∞♀)
-```
-
-**Содержательное определение:**  
-Единица направленной связи (𝟙) определяется как связь между началом смысла и концом смысла. Это базовый паттерн формы, задающий запрос ко всем связям с четко выраженной направленностью от начала к концу.
-
-**Онтологический статус:**  
-Направленность возникает из отношения между двумя частично самозамкнутыми формами. Связь получает начало и конец только в контексте этого отношения.
-
-## Ключевой принцип
-
-**Начало и конец связи определяются не как сущности, а как роли в паттернах форм:**
-
-1. Роль "начала" возникает в связях с паттерном ♂x : (♂x ⟼ x)
-2. Роль "конца" возникает в связях с паттерном x♀ : (x ⟼ x♀)
-3. Абсолютное различение начало/конец возникает только через отношение к акорню ∞
-
-## Фундаментальная теорема
-
-```
-r ⟼ v : (r ⟼ v)
-```
-
-**Содержание:** Любая связь определяется как направленное отношение между двумя другими связями, где:
-- `r` исполняет роль начала в паттерне (r ⟼ v)
-- `v` исполняет роль конца в паттерне (r ⟼ v)
-- Сама связь (r ⟼ v) возникает как новая целостность
-
-## Онтологический вывод
-
-**Начало и конец связи не предсуществуют актам связывания — они порождаются ими.**  
-В МТС нет изолированных "начал" и "концов", только паттерны форм, в которых связи исполняют соответствующие роли. Даже акорень ∞ не имеет разделения на начало и конец — это единственная связь с полным самозамыканием, из которой возникает сама возможность различения начало/конец в других связях.
-
-Это определение сохраняет онтологическую чистоту МТС: ничего кроме связей и паттернов их форм. Начало и конец — не объекты, а контекстуальные роли, определяемые через рекурсивные паттерны самозамыканий.
