--- archive/Введение в МТС.md
@@ -1,105 +1,0 @@
-# Метатеория Связей (МТС)
-## Введение: язык паттернов форм
-
-**МТС есть язык паттернов форм связей.** Это не метафора и не аналогия — это онтологический фундамент системы.
-
-В традиционных теориях символы ссылаются на внешние объекты: «стол» указывает на физический стол, «2» — на количество, «→» — на импликацию в некоторой модели. **МТС отвергает эту парадигму полностью.** Здесь нет «внешнего мира», на который ссылаются знаки. Есть только *паттерны форм связей*, которые сами являются связями.
-
----
-
-### Знак как паттерн, а не ссылка
-
-Когда мы пишем аксиому `∞ : ∞ ⟼ ∞`, символ `∞` не обозначает «бесконечность» как внешний объект. Он определяет *паттерн формы* — запрос ко всем связям, структура которых удовлетворяет условию «x = x ⟼ x». Это не ссылка, а *формальный критерий отбора* из «океана связей».
-
-Точно так же, когда мы пишем `♂x : ♂x ⟼ x`, символ `♂` не обозначает «начало» в каком-то внешнем смысле. Он определяет паттерн самозамыкания по началу: запрос ко всем связям, где начало совпадает с самой связью.
-
-**Знак в МТС — это всегда паттерн формы.** Он не указывает «туда», а задаёт правило различения форм здесь и сейчас.
-
----
-
-### Что мы различаем?
-
-Субъект (наблюдатель, интерпретатор) фиксирует только различия между состояниями. Сначала было состояние A, затем стало состояние B — это различие d(A→B). Затем различие между d(A→B) и состоянием C — новое различие d(AB→C). 
-
-**Больше субъект ничего не различает.** Он не видит «объектов», не воспринимает «свойства», не ощущает «отношения». Он фиксирует последовательность различий. 
-
-МТС делает радикальное предположение: по ту сторону субъекта находится океан связей. Каждое различие d, которое субъект фиксирует, соответствует некоторой связи в этом океане. Но субъект никогда не имеет прямого доступа к этим связям — он работает только со своими различиями.
-
----
-
-### Рекурсивные паттерны как основа
-
-Система аксиом МТС представляет собой **взаимно рекурсивную систему паттернов форм**. Каждая аксиома определяет паттерн, используя другие паттерны, и в совокупности они задают целостную онтологию.
-
-Например:
-- `∞` определяется через `∞ ⟼ ∞`
-- `♂x` определяется через `♂x ⟼ x`
-- `x♀` определяется через `x ⟼ x♀`
-- `⟼` определяется через `♂∞ ⟼ ∞♀`
-
-Это не порочный круг, а **синхронное определение системы**. Все паттерны возникают одновременно в едином акте различения. Нет «первичных» и «вторичных» понятий — есть только взаимозависимая сеть форм.
-
----
-
-### МТС как последовательность связей
-
-Сам текст МТС — это последовательность знаков. Но каждый знак в этой последовательности сам является связью, определённой паттерном формы. Поэтому:
-
-**МТС есть последовательность связей, описывающая паттерны форм связей.**
-
-Это не рефлексивный парадокс, а конструктивный принцип. Язык системы и предмет системы совпадают онтологически — оба являются связями. Разница только в роли: в одном контексте связь выступает как описание паттерна, в другом — как элемент описываемой структуры.
-
----
-
-### Как читать аксиомы
-
-Формулы в системе аксиом МТС не являются уравнениями в классическом смысле. Они задают **структурные тождества паттернов форм**. Когда мы пишем:
-
-`♂x = ♂x ⟼ x`
-
-это означает: «Паттерн формы, обозначаемый как ♂x, структурно тождественен паттерну формы (♂x ⟼ x)». Это не вычисление и не вывод — это констатация тождества форм.
-
-Ключевые символы в аксиомах:
-- `:` задаёт определение паттерна («символ слева соответствует форме справа»)
-- `=` задаёт структурную тождественность паттернов
-- `⟼` конструирует новую связь из двух существующих
-- `♂` и `♀` задают паттерны самозамыкания по началу и концу
-- `-` задаёт инверсию направленности паттерна
-- `{ }` описывает шаблоны множественных связей
-
----
-
-### Два домена: формы и данные
-
-Важно строго разделять два домена:
-1. **Домен форм связей** — где задаются паттерны, аксиомы и структурные тождества
-2. **Домен сериализации ачисел** — где связи представляются как последовательности абитов `( ) + -`
-
-Формулы аксиом принадлежат первому домену. Они описывают паттерны форм, а не данные для обработки. Символы `( ) + -` в аксиомах не являются абитами — они используются только при сериализации конкретных связей в ачисла.
-
----
-
-### Онтологическая чистота
-
-МТС принципиально избегает:
-- Внешних объектов («числа», «множества», «функции»)
-- Скрытых параметров («контекст», «интерпретация», «модель»)
-- Неявных предпосылок («допустим, что существует...»)
-
-Всё, что существует в МТС, явно задаётся паттернами форм связей. Если паттерн не определён аксиомами — его нет в системе.
-
----
-
-### С чего начинается МТС?
-
-МТС начинается не с объектов, не с аксиом, не с логики. **МТС начинается с различения.** 
-
-Первое различие: между тем, что было, и тем, что стало.  
-Второе различие: между первым различием и новым состоянием.  
-Третье различие: между вторым различием и следующим состоянием...
-
-Эта последовательность различений и есть первичная реальность для субъекта. МТС даёт язык для описания форм этих различений — язык паттернов связей.
-
-Далее следует система аксиом — рекурсивная сеть паттернов, которая формализует этот язык и делает его конструктивно управляемым. Каждая аксиома — не истинное высказывание о мире, а **конструктор различения**, задающий новый способ различать формы в океане связей.
-
-*Читая аксиомы, помните: вы работаете не с символами, ссылающимися на объекты, а с паттернами форм, задающими способы различения. Каждый символ в формуле — это уже связь, определённая своим паттерном. МТС есть самоорганизующаяся сеть паттернов, где каждый узел определяется через другие узлы в едином акте онтологического различения.*
