--- archive/Требования к изложению системы аксиом.md
@@ -1,93 +1,0 @@
-# Требования к изложению системы аксиом Метатеории связей (МТС)
-
-## 1. Онтологическая чистота
-- **Запрет внешних понятий**: Нельзя использовать термины, не выраженные в терминах связей («объект», «свойство», «внешний мир»).
-- **Самодостаточность языка**: Даже язык описания аксиом должен быть формализуем в самом языке МТС.
-- **Отсутствие онтологических теней**: Всё должно быть явной связью, никаких скрытых параметров или неявных предпосылок.
-
-## 2. Стандартизированная структура изложения каждой аксиомы
-- **Номер и название**: Четкая идентификация аксиомы.
-- **Формула**: Математическая запись в виде тождества или равенства.
-- **Смысловое определение**: Краткое онтологическое пояснение, что порождает/фиксирует аксиома.
-- **Остенсивное определение**: Графическое представление и объяснение, почему форма оператора соответствует его смыслу.
-- **Структурные свойства**: Формальные характеристики (ассоциативность, арность, симметрия).
-- **Ролевой контекст**: Как роли (ссылки/значения) возникают в акте связывания.
-- **Запреты/ограничения**: Четкое указание недопустимых интерпретаций.
-- **Следствия**: Прямые выводы и практические применения.
-
-## 3. Синтаксическая строгость
-- **Четкие правила группировки**: Однозначное указание порядка вычисления (левоассоциативность/правоассоциативность).
-- **Запрет транзитивности**: Явное указание, что из `a → b` и `b → c` не следует `a → c`.
-- **Правила использования скобок**: Четкое определение, когда и как скобки изменяют порядок группировки.
-
-## 4. Иконический характер операторов
-- **Графическая соответствие**: Форма символа должна визуально отражать его структуру (стрелка → отражает направленность).
-- **Остенсивность**: Операторы должны показывать свою структуру через форму (♂ и ♀ должны отражать замыкание).
-- **Конструктивность**: Операторы не описывают, а непосредственно конструируют бытие.
-
-## 5. Минимальность и независимость
-- **Атомарность**: Каждая аксиома вводит ровно одно новое понятие или конструкцию.
-- **Независимость**: Никакая аксиома не должна выводиться из других.
-- **Рекурсивная опора**: Аксиомы должны опираться либо на самих себя (базовый случай), либо на уже введенные аксиомы.
-
-## 6. Конструктивная направленность
-- **Акцент на порождении**: Аксиомы должны описывать, как связи конструируются, а не только их свойства.
-- **Порядок конструирования**: Четкое указание последовательности действий при построении сложных структур.
-- **Явная онтогенез**: Демонстрация, как из простых конструкций порождаются сложные.
-
-## 7. Полнота для практического применения
-- **Сериализация/десериализация**: Аксиомы должны поддерживать однозначное преобразование связей в ачисла и обратно.
-- **Вычислимость**: Возможность реализации алгоритмов обработки на основе аксиом.
-- **Примеры применения**: Конкретные демонстрации работы аксиом на практических примерах (кодирование текста, чисел, логических значений).
-
-## 8. Самореферентная согласованность
-- **Метауровень**: Возможность описания самой системы МТС на языке МТС.
-- **Отсутствие парадоксов**: Гарантии от самореферентных противоречий при работе с рекурсивными конструкциями.
-- **Границы применимости**: Четкое указание, где заканчивается область действия аксиом.
-
-## 9. Избегание философской неопределенности
-- **Замена интуиции формализмом**: Отказ от эвристических или поэтических формулировок в пользу строгих конструктивных определений.
-- **Отказ от антропоморфизма**: Запрет терминов вроде «мы», «понимаем», «значит» в базовых аксиомах.
-- **Точность терминов**: Каждый термин должен быть строго связан с формальной конструкцией.
-
-## 10. Логическая последовательность изложения
-- **От простого к сложному**: Аксиомы должны идти в порядке возрастания сложности.
-- **Зависимости**: Каждая аксиома может ссылаться только на предыдущие.
-- **Иерархия**: Четкое разделение базовых (универсальных) и производных аксиом.
-
-Соблюдение этих требований позволит создать строгую, непротиворечивую и практичную систему аксиом МТС, которая будет одновременно философски глубокой и технически применимой.
-
----
-
-
-Замечания для чистовика МТС:
-
-1. У каждого вводимого символа в формальной нотации МТС, есть своя аксиома с названием
-2. Название аксиомы вводит название символа как смысл
-3. Содержательно-онтологическое определение должно быть последовательным прочтением формулы, т.е. названия аксиом символов
-4. Аксиомы должны следовать логической последовательности от фундаментальных к производным
-5. Использовать определение ':') для введения новых символов
-6. Использовать равенство '=' для аксиомы равенства связей по смыслу
-7. Заменить "роль" на "смысл"
-8. Заменить "ссылка/значение" на "начало/конец"
-9. Заменить "различие" на "смысл"
-10. Уточнить терминологию "оператор" как "оператор формальной нотации МТС", что есть оператор?
-11. Метаязык, т.е. формальная нотация МТС, которая вводит символы для описания формы связей и их свойств
-12. ачисло это уникальный идентификатор связи, который можно использовать как ссылку на связь
-13. использовать ':' как знак аксиомы определения, для определениях знаков/имён/смыслов новой связи
-14. аксиома инверсии направления связи вводит знак ¬ смены направления связи (не использовать минус -)
-15. '(' : ♂∞ → ∞
-16. ')' : ∞ → ∞♀
-17. '1' : ♂∞ → ∞♀
-18. '0' : ∞♀ → ♂∞
-
-Если мы определяем ♂ как унарный оператор и форму самозамыкания связи, с самозамкнутым началом, т.е. отсутствием начала, то это по сути безначальная связь, т.е. у неё есть только конец, и её смысл в том что это конец связи.
-
-Если мы определяем ♀ как унарный оператор и форму самозамыкания связи, с самозамкнутым концом, т.е. отсутствием конца, то это по сути бесконечная связь, т.е. у неё есть только начало, и её смысл в том что это начало связи.
-
-Теперь, можно определить, что эти операторы, при применении к связи, возвращают либо начало либо конец связи, т.е.:
-♂ab = b
-ab♀ = a
-
-Если так то теперь у нас получается возможность определить знак инверсии связи:
-¬ab : ♂ab → ab♀
